Rename js file.
